Fixed PotionInfo build failing

+++ src/main/java/me/zeroeightsix/kami/util/PotionInfo.kt
@@ -10,8 +10,8 @@ data class PotionInfo(val name: String, val amplifier: Int, val potionEffect: Po
      * (min:secs)
      */
     fun formattedTimeLeft(): String {
-        val tps = LagCompensator.tickRate.toInt()
-        val compTps = if (tps != 0) tps else 20
+        val tps = LagCompensator.tickRate.toLong()
+        val compTps = if (tps != 0L) tps else 20L
         val compensatedDuration: Long = potionEffect.duration / compTps
         val min = TimeUnit.SECONDS.toMinutes(compensatedDuration)
         val secs = TimeUnit.SECONDS.toSeconds(compensatedDuration) - min * 60
